Conflicts:
	.DS_Store
	config.go
	config/LICENSE
	config/global.yaml
	config/local_peerconfig.json
	config/pbft.yaml
	consensus/pbft/pbftprotocal.go
	hpc/nodeAPI.go
	hpc/transactionAPI.go
	main.go
	manager/backend.go
	manager/protocolManager.go
	p2p/gRPCPeerManager.go
	p2p/node.go
	p2p/node_test.go
	p2p/peer.go
	p2p/peerInfo.go
	p2p/peerManager.go
	p2p/peerPool.go
	p2p/peerPool_test.go
	p2p/peer_test.go
	p2p/peermessage/message.pb.go
	p2p/peermessage/message.proto
	p2p/transport/transport.go
	scripts/LICENSE-TEMP
	scripts/license_generator.go
	scripts/local.sh

+++ core/blockpool/block_pool.go
@@ -1,7 +1,5 @@
-// implementblock pool
-// author: Lizhong kuang
-// date: 2016-08-29
-// last modified:2016-09-01
+//Hyperchain License
+//Copyright (C) 2016 The Hyperchain Authors.
 package blockpool
 
 import (
@@ -26,10 +24,10 @@ import (
 	"sync/atomic"
 	"time"
 	"sort"
+	"errors"
 )
 
 var (
-	public_batch hyperdb.Batch
 	log          *logging.Logger // package-level logger
 )
 
@@ -55,16 +53,21 @@ type BlockPool struct {
 	eventMux            *event.TypeMux
 	stateLock           sync.Mutex
 	wg                  sync.WaitGroup // for shutdown sync
-	lastValidationState common.Hash
+	lastValidationState atomic.Value
 	blockCache          *common.Cache
 	validationQueue     *common.Cache
 	queue               *common.Cache
 }
 
 func NewBlockPool(eventMux *event.TypeMux, consenter consensus.Consenter) *BlockPool {
-	blockcache, _ := common.NewCache()
-	queue, _ := common.NewCache()
-	validationqueue, _ := common.NewCache()
+	var err error
+	blockcache, err := common.NewCache()
+	if err != nil {return nil}
+	queue, err := common.NewCache()
+	if err != nil {return nil}
+	validationqueue, err := common.NewCache()
+	if err != nil {return nil}
+
 	pool := &BlockPool{
 		eventMux:        eventMux,
 		consenter:       consenter,
@@ -72,13 +75,15 @@ func NewBlockPool(eventMux *event.TypeMux, consenter consensus.Consenter) *Block
 		validationQueue: validationqueue,
 		blockCache:      blockcache,
 	}
-
 	currentChain := core.GetChainCopy()
 	pool.demandNumber = currentChain.Height + 1
 	pool.demandSeqNo = currentChain.Height + 1
-	db, _ := hyperdb.GetLDBDatabase()
-	blk, _ := core.GetBlock(db, currentChain.LatestBlockHash)
-	pool.lastValidationState = common.BytesToHash(blk.MerkleRoot)
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {return nil}
+
+	blk, err := core.GetBlock(db, currentChain.LatestBlockHash)
+	if err != nil {return nil}
+	pool.lastValidationState.Store(common.BytesToHash(blk.MerkleRoot))
 
 	log.Noticef("Block pool Initialize demandNumber :%d, demandseqNo: %d\n", pool.demandNumber, pool.demandSeqNo)
 	return pool
@@ -115,6 +120,9 @@ func (pool *BlockPool) Validate(validationEvent event.ExeTxsEvent, commonHash cr
 		if _, success := pool.PreProcess(validationEvent, commonHash, encryption, peerManager); success {
 			atomic.AddUint64(&pool.demandSeqNo, 1)
 			log.Notice("Current demandSeqNo is, ", pool.demandSeqNo)
+		} else {
+			log.Error("PreProcess Failed")
+			return
 		}
 		judge := func(key interface{}, iterKey interface{}) bool {
 			id := key.(uint64)
@@ -134,6 +142,9 @@ func (pool *BlockPool) Validate(validationEvent event.ExeTxsEvent, commonHash cr
 					pool.validationQueue.Remove(i)
 					atomic.AddUint64(&pool.demandSeqNo, 1)
 					log.Notice("Current demandSeqNo is, ", pool.demandSeqNo)
+				} else {
+					log.Error("PreProcess Failed")
+					return
 				}
 			} else {
 				break
@@ -161,10 +172,15 @@ func (pool *BlockPool) PreProcess(validationEvent event.ExeTxsEvent, commonHash 
 		sort.Ints(index)
 		count := 0
 		set := validationEvent.Transactions
-		for i := range index {
-			i = i-count
-			set = append(set[:i-1], set[i+1:]...)
-			count++
+		for _, idx := range index {
+			idx = idx-count
+			if idx == 0 {
+				set = set[1:]
+				count++
+			} else {
+				set = append(set[:idx - 1], set[idx + 1:]...)
+				count++
+			}
 		}
 		validTxSet = set
 	} else {
@@ -172,6 +188,7 @@ func (pool *BlockPool) PreProcess(validationEvent event.ExeTxsEvent, commonHash 
 	}
 	err, _, merkleRoot, txRoot, receiptRoot, validTxSet, invalidTxSet := pool.ProcessBlockInVm(validTxSet, invalidTxSet, validationEvent.SeqNo)
 	if err != nil {
+		log.Error("ProcessBlock Failed!, block number: ", validationEvent.SeqNo)
 		return err, false
 	}
 	hash := commonHash.Hash([]interface{}{
@@ -228,19 +245,24 @@ func (pool *BlockPool) CheckSign(txs []*types.Transaction, commonHash crypto.Com
 	// (1) check signature for each transaction
 	var wg sync.WaitGroup
 	var index []int
-	for i, tx := range txs {
+	var mu sync.Mutex
+	for i := range txs {
 		wg.Add(1)
-		go func(tx *types.Transaction){
+		go func(i int){
+			tx := txs[i]
 			if !tx.ValidateSign(encryption, commonHash) {
 				log.Notice("Validation, found invalid signature, send from :", tx.Id)
 				invalidTxSet = append(invalidTxSet, &types.InvalidTransactionRecord{
 					Tx:      tx,
 					ErrType: types.InvalidTransactionRecord_SIGFAILED,
+					ErrMsg:  []byte("Invalid signature"),
 				})
+				mu.Lock()
 				index = append(index, i)
+				mu.Unlock()
 			}
 			wg.Done()
-		}(tx)
+		}(i)
 	}
 	wg.Wait()
 	return invalidTxSet, index
@@ -258,36 +280,67 @@ func (pool *BlockPool) ProcessBlockInVm(txs []*types.Transaction, invalidTxs []*
 	if err != nil {
 		return err, nil, nil, nil, nil, nil, invalidTxs
 	}
-	txTrie, _ := trie.New(common.Hash{}, db)
-	receiptTrie, _ := trie.New(common.Hash{}, db)
-	statedb, err := state.New(pool.lastValidationState, db)
-	if err != nil {
-		log.Error("New StateDB ERROR")
-		return err, nil, nil, nil, nil, nil, invalidTxs
+	txTrie, err := trie.New(common.Hash{}, db)
+	if err != nil {return err, nil, nil, nil, nil, nil, invalidTxs}
+	receiptTrie, err := trie.New(common.Hash{}, db)
+	if err != nil {return err, nil, nil, nil, nil, nil, invalidTxs}
+	v := pool.lastValidationState.Load()
+	initStatus, ok := v.(common.Hash)
+	if ok == false {
+		return errors.New("Get StateDB Status Failed!"), nil, nil, nil, nil, nil, invalidTxs
 	}
+	statedb, err := state.New(initStatus, db)
+
+	if err != nil {return err, nil, nil, nil, nil, nil, invalidTxs}
 	env["currentNumber"] = strconv.FormatUint(seqNo, 10)
 	env["currentGasLimit"] = "10000000"
 	vmenv := core.NewEnvFromMap(core.RuleSet{params.MainNetHomesteadBlock, params.MainNetDAOForkBlock, true}, statedb, env)
 
-	public_batch = db.NewBatch()
+	public_batch := db.NewBatch()
 	for i, tx := range txs {
 		statedb.StartRecord(tx.GetTransactionHash(), common.Hash{}, i)
 		receipt, _, _, err := core.ExecTransaction(*tx, vmenv)
-		if err != nil && core.IsValueTransferErr(err) {
+		if err != nil{
+			var errType types.InvalidTransactionRecord_ErrType
+			if core.IsValueTransferErr(err) {
+				errType = types.InvalidTransactionRecord_OUTOFBALANCE
+			} else if core.IsExecContractErr(err) {
+				tmp := err.(*core.ExecContractError)
+				if tmp.GetType() == 0 {
+					errType = types.InvalidTransactionRecord_DEPLOY_CONTRACT_FAILED
+				} else if tmp.GetType() == 1{
+					errType = types.InvalidTransactionRecord_INVOKE_CONTRACT_FAILED
+				} else {
+					// For extension
+				}
+			} else {
+				// For extension
+			}
 			invalidTxs = append(invalidTxs, &types.InvalidTransactionRecord{
 				Tx:      tx,
-				ErrType: types.InvalidTransactionRecord_OUTOFBALANCE,
+				ErrType: errType,
+				ErrMsg:  []byte(err.Error()),
 			})
 			continue
 		}
 		// save to DB
-		txValue, _ := proto.Marshal(tx)
+		txValue, err := proto.Marshal(tx)
+		if err != nil {
+			log.Error("Invalid Transaction struct to marshal! error msg, ", err.Error())
+			return err, nil, nil, nil, nil, nil, invalidTxs
+		}
 		if err := public_batch.Put(append(core.TransactionPrefix, tx.GetTransactionHash().Bytes()...), txValue); err != nil {
+			log.Error("Put tx data into database failed! error msg, ", err.Error())
 			return err, nil, nil, nil, nil, nil, invalidTxs
 		}
 
-		receiptValue, _ := proto.Marshal(receipt)
+		receiptValue, err := proto.Marshal(receipt)
+		if err != nil {
+			log.Error("Invalid receipt struct to marshal! error msg, ", err.Error())
+			return err, nil, nil, nil, nil, nil, invalidTxs
+		}
 		if err := public_batch.Put(append(core.ReceiptsPrefix, receipt.TxHash...), receiptValue); err != nil {
+			log.Error("Put receipt data into database failed! error msg, ", err.Error())
 			return err, nil, nil, nil, nil, nil, invalidTxs
 		}
 		// set temporarily
@@ -296,7 +349,11 @@ func (pool *BlockPool) ProcessBlockInVm(txs []*types.Transaction, invalidTxs []*
 			BlockIndex: seqNo,
 			Index:      int64(i),
 		}
-		metaValue, _ := proto.Marshal(meta)
+		metaValue, err := proto.Marshal(meta)
+		if err != nil {
+			log.Error("Invalid txmeta struct to marshal! error msg, ", err.Error())
+			return err, nil, nil, nil, nil, nil, invalidTxs
+		}
 		if err := public_batch.Put(append(tx.GetTransactionHash().Bytes(), core.TxMetaSuffix...), metaValue); err != nil {
 			log.Error("Put txmeta into database failed! error msg, ", err.Error())
 			return err, nil, nil, nil, nil, nil, invalidTxs
@@ -306,11 +363,15 @@ func (pool *BlockPool) ProcessBlockInVm(txs []*types.Transaction, invalidTxs []*
 		receiptTrie.Update(append(core.ReceiptsPrefix, receipt.TxHash...), receiptValue)
 		validtxs = append(validtxs, tx)
 	}
-	root, _ := statedb.Commit()
+	root, err := statedb.Commit()
+	if err != nil {
+		log.Error("Commit state db failed! error msg, ", err.Error())
+		return err, nil, nil, nil, nil, nil, invalidTxs
+	}
 	merkleRoot := root.Bytes()
 	txRoot := txTrie.Hash().Bytes()
 	receiptRoot := receiptTrie.Hash().Bytes()
-	pool.lastValidationState = root
+	pool.lastValidationState.Store(root)
 	go public_batch.Write()
 	return nil, nil, merkleRoot, txRoot, receiptRoot, validtxs, invalidTxs
 }
@@ -362,17 +423,9 @@ func (pool *BlockPool) CommitBlock(ev event.CommitOrRollbackBlockEvent, commonHa
 				peerManager.SendMsgToPeers(payload, peers, recovery.Message_INVALIDRESP)
 			}
 		}
-	} else {
-		// this branch will never activated
-		// instead of send an CommitOrRollbackBlockEvent with `false` flag, PBFT send a `viewchange` or `self recovery`
-		// message to handle this issue
-		// TODO remove this branch
-		db, _ := hyperdb.GetLDBDatabase()
-		for _, t := range record.InvalidTxs {
-			db.Delete(append(core.TransactionPrefix, t.Tx.GetTransactionHash().Bytes()...))
-			db.Delete(append(core.ReceiptsPrefix, t.Tx.GetTransactionHash().Bytes()...))
-		}
 	}
+	// instead of send an CommitOrRollbackBlockEvent with `false` flag, PBFT send a `viewchange` or `self recovery`
+	// message to handle this issue
 	pool.blockCache.Remove(ev.Hash)
 }
 
@@ -421,7 +474,11 @@ func WriteBlock(block *types.Block, commonHash crypto.CommonHash, vid uint64, pr
 	log.Info("block number is ", block.Number)
 	core.UpdateChain(block, false)
 
-	db, _ := hyperdb.GetLDBDatabase()
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		log.Error("Get Database Instance Failed! error msg,", err.Error())
+		return
+	}
 	// for primary node, check whether vid equal to block's number
 	if primary && vid != block.Number {
 		log.Info("Replace invalid txmeta data, block number:", block.Number)
@@ -431,7 +488,11 @@ func WriteBlock(block *types.Block, commonHash crypto.CommonHash, vid uint64, pr
 				BlockIndex: block.Number,
 				Index:      int64(i),
 			}
-			metaValue, _ := proto.Marshal(meta)
+			metaValue, err := proto.Marshal(meta)
+			if err != nil {
+				log.Error("Invalid txmeta sturct, marshal failed! error msg,", err.Error())
+				return
+			}
 			if err := batch.Put(append(tx.GetTransactionHash().Bytes(), core.TxMetaSuffix...), metaValue); err != nil {
 				log.Error("Put txmeta into database failed! error msg, ", err.Error())
 				return
@@ -440,9 +501,10 @@ func WriteBlock(block *types.Block, commonHash crypto.CommonHash, vid uint64, pr
 		batch.Write()
 	}
 
-	err := core.PutBlockTx(db, commonHash, block.BlockHash, block)
+	err = core.PutBlockTx(db, commonHash, block.BlockHash, block)
 	if err != nil {
 		log.Error("Put block into database failed! error msg, ", err.Error())
+		return
 	}
 
 	if block.Number%10 == 0 && block.Number != 0 {
@@ -463,7 +525,11 @@ func (pool *BlockPool) StoreInvalidResp(ev event.RespInvalidTxsEvent) {
 	}
 	// save to db
 	log.Notice("invalidTx", common.BytesToHash(invalidTx.Tx.TransactionHash).Hex())
-	db, _ := hyperdb.GetLDBDatabase()
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		log.Error("Get Database Instance Failed! error msg,", err.Error())
+		return
+	}
 	db.Put(append(core.InvalidTransactionPrefix, invalidTx.Tx.TransactionHash...), ev.Payload)
 }
 
@@ -475,48 +541,38 @@ func (pool *BlockPool) ResetStatus(ev event.VCResetEvent) {
 	atomic.StoreUint64(&pool.demandSeqNo, ev.SeqNo)
 	atomic.StoreUint64(&pool.maxSeqNo, ev.SeqNo-1)
 
-	db, _ := hyperdb.GetLDBDatabase()
-	block, _ := core.GetBlockByNumber(db, ev.SeqNo-1)
-	pool.lastValidationState = common.BytesToHash(block.MerkleRoot)
-	// 2. Delete Invalid Stuff
-	for i := ev.SeqNo; i < tmpDemandNumber; i += 1 {
-		// delete tx, txmeta and receipt
-		block, err := core.GetBlockByNumber(db, i)
-		if err != nil {
-			log.Errorf("ViewChange, miss block %d ,error msg %s", i, err.Error())
-		}
-
-		for _, tx := range block.Transactions {
-			if err := db.Delete(append(core.TransactionPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
-				log.Errorf("ViewChange, delete useless tx in block %d failed, error msg %s", i, err.Error())
-			}
-			if err := db.Delete(append(core.ReceiptsPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
-				log.Errorf("ViewChange, delete useless receipt in block %d failed, error msg %s", i, err.Error())
-			}
-			if err := db.Delete(append(tx.GetTransactionHash().Bytes(), core.TxMetaSuffix...)); err != nil {
-				log.Errorf("ViewChange, delete useless txmeta in block %d failed, error msg %s", i, err.Error())
-			}
-		}
-		// delete block
-		if err := core.DeleteBlockByNum(db, i); err != nil {
-			log.Errorf("ViewChange, delete useless block %d failed, error msg %s", i, err.Error())
-		}
-
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		log.Error("Get Database Instance Failed! error msg,", err.Error())
+		return
 	}
+
+
+	block, err := core.GetBlockByNumber(db, ev.SeqNo-1)
+	if err != nil {
+		return
+	}
+	pool.lastValidationState.Store(common.BytesToHash(block.MerkleRoot))
+	// 2. Delete Invalid Stuff
+	pool.RemoveData(ev.SeqNo, tmpDemandNumber)
+
 	// 3. Delete from blockcache
 	keys := pool.blockCache.Keys()
 	for _, key := range keys {
 		ret, _ := pool.blockCache.Get(key)
+		if ret == nil {
+			continue
+		}
 		record := ret.(BlockRecord)
 		for i, tx := range record.ValidTxs {
 			if err := db.Delete(append(core.TransactionPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
-				log.Errorf("ViewChange, delete useless tx in block %d failed, error msg %s", i, err.Error())
+				log.Errorf("ViewChange, delete useless tx in cache %d failed, error msg %s", i, err.Error())
 			}
 			if err := db.Delete(append(core.ReceiptsPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
-				log.Errorf("ViewChange, delete useless receipt in block %d failed, error msg %s", i, err.Error())
+				log.Errorf("ViewChange, delete useless receipt in cache %d failed, error msg %s", i, err.Error())
 			}
 			if err := db.Delete(append(tx.GetTransactionHash().Bytes(), core.TxMetaSuffix...)); err != nil {
-				log.Errorf("ViewChange, delete useless txmeta in block %d failed, error msg %s", i, err.Error())
+				log.Errorf("ViewChange, delete useless txmeta in cache %d failed, error msg %s", i, err.Error())
 			}
 		}
 	}
@@ -524,5 +580,118 @@ func (pool *BlockPool) ResetStatus(ev event.VCResetEvent) {
 	// 4. Reset chain
 	isGenesis := (block.Number == 0)
 	core.UpdateChain(block, isGenesis)
-
 }
+func (pool *BlockPool) RunInSandBox(tx *types.Transaction) error {
+	// TODO add block number to specify the initial status
+	var env = make(map[string]string)
+	fakeBlockNumber := core.GetHeightOfChain()
+	env["currentNumber"] = strconv.FormatUint(fakeBlockNumber, 10)
+	env["currentGasLimit"] = "10000000"
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		return err
+	}
+	v := pool.lastValidationState.Load()
+	initStatus, ok := v.(common.Hash)
+	if ok == false {
+		return errors.New("Get StateDB Status Failed!")
+	}
+	statedb, err := state.New(initStatus, db)
+	if err != nil {
+		return err
+	}
+	sandBox := core.NewEnvFromMap(core.RuleSet{params.MainNetHomesteadBlock, params.MainNetDAOForkBlock, true}, statedb, env)
+	receipt, _, _, err := core.ExecTransaction(*tx, sandBox)
+	if err != nil{
+		var errType types.InvalidTransactionRecord_ErrType
+		if core.IsValueTransferErr(err) {
+			errType = types.InvalidTransactionRecord_OUTOFBALANCE
+		} else if core.IsExecContractErr(err) {
+			tmp := err.(*core.ExecContractError)
+			if tmp.GetType() == 0 {
+				errType = types.InvalidTransactionRecord_DEPLOY_CONTRACT_FAILED
+			} else if tmp.GetType() == 1{
+				errType = types.InvalidTransactionRecord_INVOKE_CONTRACT_FAILED
+			} else {
+				// For extension
+			}
+		} else {
+			// For extension
+		}
+		t :=  &types.InvalidTransactionRecord{
+			Tx:      tx,
+			ErrType: errType,
+			ErrMsg:  []byte(err.Error()),
+		}
+		payload, err := proto.Marshal(t)
+		if err != nil {
+			log.Error("Marshal tx error")
+			return nil
+		}
+		pool.StoreInvalidResp(event.RespInvalidTxsEvent{
+			Payload: payload,
+		})
+		return nil
+	} else {
+		receiptValue, err := proto.Marshal(receipt)
+		if err != nil {
+			log.Error("Invalid receipt struct to marshal! error msg, ", err.Error())
+			return err
+		}
+		if err := db.Put(append(core.ReceiptsPrefix, receipt.TxHash...), receiptValue); err != nil {
+			log.Error("Put receipt data into database failed! error msg, ", err.Error())
+			return err
+		}
+		return nil
+	}
+}
+
+func (pool *BlockPool) RemoveData(from, to uint64) {
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		log.Error("Get Database Instance Failed! error msg,", err.Error())
+		return
+	}
+
+	// delete tx, txmeta and receipt
+	for i := from; i < to; i += 1 {
+		block, err := core.GetBlockByNumber(db, i)
+		if err != nil {
+			log.Errorf("miss block %d ,error msg %s", i, err.Error())
+			continue
+		}
+
+		for _, tx := range block.Transactions {
+			if err := db.Delete(append(core.TransactionPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
+				log.Errorf("delete useless tx in block %d failed, error msg %s", i, err.Error())
+			}
+			if err := db.Delete(append(core.ReceiptsPrefix, tx.GetTransactionHash().Bytes()...)); err != nil {
+				log.Errorf("delete useless receipt in block %d failed, error msg %s", i, err.Error())
+			}
+			if err := db.Delete(append(tx.GetTransactionHash().Bytes(), core.TxMetaSuffix...)); err != nil {
+				log.Errorf("delete useless txmeta in block %d failed, error msg %s", i, err.Error())
+			}
+		}
+		// delete block
+		if err := core.DeleteBlockByNum(db, i); err != nil {
+			log.Errorf("ViewChange, delete useless block %d failed, error msg %s", i, err.Error())
+		}
+	}
+}
+
+func (pool *BlockPool) CutdownBlock(number uint64) {
+	pool.RemoveData(number, number + 1)
+	db, err := hyperdb.GetLDBDatabase()
+	if err != nil {
+		log.Error("Get Database Instance Failed! error msg,", err.Error())
+		return
+	}
+	block, err := core.GetBlockByNumber(db, number - 1)
+	if err != nil {
+		log.Errorf("miss block %d ,error msg %s", number - 1, err.Error())
+		return
+	}
+	pool.lastValidationState.Store(common.BytesToHash(block.MerkleRoot))
+}
+
+
